Remove zkp.js and update zkp.ts with proper implementation
