+++ src/lib/zkp.ts
@@ -1,21 +1,115 @@
-/**
- * Zero-Knowledge Proof (ZKP) utilities
- * 
- * This module provides functions for generating and verifying ZKP proofs.
- * It implements a password-based authentication system where the password
- * is never sent to the server.
- */
+// ZKP (Zero-Knowledge Proof) Library
+import * as snarkjs from 'snarkjs';
+import * as fs from 'fs';
+import * as path from 'path';
+import * as crypto from 'crypto';
 
-interface CredentialInput {
-  username: string;
-  password: string;
-  salt: string;
+// Define interfaces for the ZKP system
+export interface Proof {
+  pi_a: string[] | number[];
+  pi_b: (string[] | number[])[];
+  pi_c: string[] | number[];
+  protocol: string;
+  curve: string;
+  tampered?: boolean;
 }
 
-interface ProofVerificationInput {
-  proof: any;
-  publicSignals: any;
-  publicKey: any;
+export interface ZKPResult {
+  proof: Proof;
+  publicSignals: string[];
+}
+
+/**
+ * Generate a public key from credentials
+ * @param username - The username
+ * @param password - The password
+ * @param salt - The salt
+ * @returns The public key
+ */
+export async function generatePublicKey(username: string, password: string, salt: string): Promise<string> {
+  // Combine the inputs
+  const combined = `${username}:${password}:${salt}`;
+
+  // Create a SHA-256 hash
+  return crypto.createHash('sha256').update(combined).digest('hex');
+}
+
+/**
+ * Generate a ZKP proof for authentication
+ * @param username - The username
+ * @param password - The password
+ * @param salt - The salt
+ * @returns The proof and public signals
+ */
+export async function generateProof(username: string, password: string, salt: string): Promise<ZKPResult> {
+  try {
+    // Get the circuit paths
+    const circuitWasmPath = path.join(process.cwd(), 'circuits/zkp_auth/simple_auth_output/simple_auth_js/simple_auth.wasm');
+    const zkeyPath = path.join(process.cwd(), 'circuits/zkp_auth/simple_auth_output/simple_auth_final.zkey');
+
+    // Ensure the circuit files exist
+    if (!fs.existsSync(circuitWasmPath) || !fs.existsSync(zkeyPath)) {
+      throw new Error('Circuit files not found. Please compile the circuits first.');
+    }
+
+    // Create the witness input
+    const input = {
+      username: stringToHex(username),
+      password: stringToHex(password),
+      salt: stringToHex(salt),
+      timestamp: Date.now().toString()
+    };
+
+    // Generate the proof
+    const { proof, publicSignals } = await snarkjs.groth16.fullProve(
+      input,
+      circuitWasmPath,
+      zkeyPath
+    );
+
+    return { proof, publicSignals } as ZKPResult;
+  } catch (error) {
+    console.error('Error generating proof:', error);
+    throw error;
+  }
+}
+
+/**
+ * Verify a ZKP proof
+ * @param proof - The proof
+ * @param publicSignals - The public signals
+ * @returns Whether the proof is valid
+ */
+export async function verifyProof(proof: Proof, publicSignals: string[]): Promise<boolean> {
+  try {
+    // Get the verification key
+    const vKeyPath = path.join(process.cwd(), 'circuits/zkp_auth/simple_auth_output/verification_key.json');
+
+    // Ensure the verification key exists
+    if (!fs.existsSync(vKeyPath)) {
+      throw new Error('Verification key not found. Please compile the circuits first.');
+    }
+
+    // Load the verification key
+    const vKey = JSON.parse(fs.readFileSync(vKeyPath, 'utf8'));
+
+    // Verify the proof
+    const isValid = await snarkjs.groth16.verify(vKey, publicSignals, proof);
+
+    return isValid;
+  } catch (error) {
+    console.error('Error verifying proof:', error);
+    return false;
+  }
+}
+
+/**
+ * Convert a string to hex
+ * @param str - The string to convert
+ * @returns The hex representation
+ */
+function stringToHex(str: string): string {
+  return '0x' + Buffer.from(str, 'utf8').toString('hex');
 }
 
 /**
@@ -23,77 +117,5 @@ interface ProofVerificationInput {
  * @returns A random salt string
  */
 export function generateSalt(): string {
-  return Math.random().toString(36).substring(2, 15) + 
-         Math.random().toString(36).substring(2, 15);
-}
-
-/**
- * Generate a public key from a username, password, and salt
- * @param input The credential input
- * @returns The public key
- */
-export async function generatePublicKey(input: CredentialInput): Promise<string> {
-  // In a real implementation, this would use a proper ZKP library
-  // For now, we'll use a simple hash-based approach for demonstration
-
-  // Combine the inputs
-  const combined = `${input.username}:${input.password}:${input.salt}`;
-  
-  // Convert to a hash (in a real implementation, this would be more complex)
-  const encoder = new TextEncoder();
-  const data = encoder.encode(combined);
-  const hashBuffer = await crypto.subtle.digest('SHA-256', data);
-  
-  // Convert to a string
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-  const hashHex = hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
-  
-  return hashHex;
-}
-
-/**
- * Generate a zero-knowledge proof for authentication
- * @param input The credential input
- * @returns The proof and public signals
- */
-export async function generateProof(input: CredentialInput): Promise<{
-  proof: any;
-  publicSignals: any;
-}> {
-  // In a real implementation, this would use a proper ZKP library
-  // For testing purposes, we'll create a simplified version
-  
-  // Generate a mock proof
-  const proof = {
-    pi_a: [`${input.username}_proof_a`, "2", "3"],
-    pi_b: [["4", "5"], ["6", "7"], ["8", "9"]],
-    pi_c: [`${input.password.length}_proof_c`, "11", "12"],
-    protocol: "groth16"
-  };
-  
-  // Generate mock public signals
-  const publicSignals = [
-    await generatePublicKey(input),
-    `${Date.now()}`
-  ];
-  
-  return { proof, publicSignals };
-}
-
-/**
- * Verify a zero-knowledge proof
- * @param input The proof verification input
- * @returns Whether the proof is valid
- */
-export async function verifyProof(input: ProofVerificationInput): Promise<boolean> {
-  // In a real implementation, this would use a proper ZKP library
-  // For testing purposes, we'll return true for valid test users
-  
-  try {
-    // Check if the first public signal matches the public key
-    return input.publicSignals[0] === input.publicKey;
-  } catch (error) {
-    console.error('Error verifying proof:', error);
-    return false;
-  }
+  return crypto.randomBytes(16).toString('hex');
 }
